Move Strategy to types.tsx & create handleSubmit.
