+++ src/store/types.tsx
@@ -1,0 +1,10 @@
+type Frequency = "Annually" | "Monthly" | "Fortnightly" | "Weekly" | "Daily"
+
+export interface Strategy {
+  initialDeposit: number,
+  regularDeposit: number,
+  depositFrequency: Frequency,
+  compoundFrequency: Frequency,
+  numberOfYears: number,
+  annualInterestRate: number
+}
